requirements.txt
aiogram>=3.4.1,<4.0.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0

